+++ wdfwd/tests/test_db.py
@@ -94,11 +94,11 @@ def test_db_data_dates():
     dcfg2 = write_eloa_cfg(dcfg)
     with db.Connector(dcfg2) as con:
         dates = db.get_data_dates(con, False)
-        assert dates == ['2015/11/23', '2015/11/24', '2015/11/25']
+        assert dates == ['2015-11-23', '2015-11-24', '2015-11-25']
 
 
 def test_db_table_rowcnt_by_date():
     dcfg2 = write_eloa_cfg(dcfg)
     with db.Connector(dcfg2) as con:
-        n_rows = db.table_rowcnt_by_date(con, 'CharStateLog_TBL', '2015/11/23')
+        n_rows = db.table_rowcnt_by_date(con, 'CharStateLog_TBL', '2015-11-23')
         assert n_rows == 375476